2024 D14P1
Heh, so I got tripped up during the seconds-iteration as the robots weren't ending up where they should. It's because the robots would move themselves into slots that were yet to be iterated to in the same second. So if a robot had a (1, 1) velocity, they could be moved several times in the same second.