2024 D14P2
Pretty lazy solution: the robots will arrange themselves into a Christmas tree within a perfect-rectangle frame. As such, i figure if that if you detect a perfect corner for at least six slots then it's extremely likely you've found it. It then just prints the grid to a file as a manual check and voila.